Faster Benchmarking
For dequeuing and preceding time tests, the benchmark will fill a runqueue just before each test which will cause a significant slowdown in performing the benchmark. This was prevalent in the OrderedArray runqueue which is the slowest in enqueuing data.

Now the benchmark creates a runqueue with data up to the size tested for each series of tests. Example: A filled runqueue will be created for a dequeue test, but NOT for every iteration of that test.